requirements.txt
cachetools==5.3.2
exceptiongroup==1.2.0
google-auth==2.26.2
graalpy-virtualenv==0.0.1
hpy==0.9.0rc2
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pytest==7.4.4
rsa==4.9
tomli==2.0.1
